Extract Markerable base class for ListItem and MarkupSection
