+++ src/js/models/list-item.js
@@ -1,8 +1,8 @@
-import Section from './markup-section';
+import Markerable from './_markerable';
 
 export const LIST_ITEM_TYPE = 'list-item';
 
-export default class ListItem extends Section {
+export default class ListItem extends Markerable {
   constructor(tagName, markers=[]) {
     super(tagName, markers);
     this.type = LIST_ITEM_TYPE;
@@ -12,15 +12,16 @@ export default class ListItem extends Section {
     // FIXME need to check if we are going to split into two list items
     // or a list item and a new markup section:
     const isLastItem = !this.next;
-    const createNewSection =
-      (marker.isEmpty && offset === 0) && isLastItem;
+    const createNewSection = (!marker && offset === 0 && isLastItem);
 
     let [beforeSection, afterSection] = [
       this.builder.createListItem(),
-      createNewSection ? this.builder.createMarkupSection('p') : this.builder.createListItem()
+      createNewSection ? this.builder.createMarkupSection() :
+                         this.builder.createListItem()
     ];
 
-    return this._redistributeMarkers(beforeSection, afterSection, marker, offset);
+    return this._redistributeMarkers(
+      beforeSection, afterSection, marker, offset);
   }
 
   splitIntoSections() {
